ai-resume-tailor/requirements.txt
streamlit
openai
python-docx
docx2pdf
